Add sauron-prio-adjust-functions hook.
This hook makes it possible to add functions that will modify the event
priority as it is added. Each function in the hook can return an integer
value which will be added to the priority of the event, promoting or
demoting it as appropriate.

This can, for example, be used to increase the priority of all events
when the emacs frame is not focused.