initial import of sauron-zeroconf
This package implements logging of zeroconf events to sauron.

It connects via the built-in zeroconf.el

If the avahi service-type is located those services will be added to the
internal list to provide up-to-date mapping.

You can filter the logging using "sauron-zeroconf-filter-function"

Selecting the item in the sauron log will open a buffer displaing the
details.  URLs will be created for _http._tcp, _http-alt._tcp, and
_https._tcp services.